Dynamic AI in Weather Forecast Prediction
We have chosen localized & populous urban sites: Haridwar, Delhi and Mathura for this research project, along the Yamuna river bank for a pervasive AI based algorithm which would apply to the variations in a river with different dynamics at different junctures of its flow; from the origin at the Yamunotri, its downward curve till Delhi and its sharp turn thereafter till it merges into the Ganges as a tributary at the banks of Varanasi. 

We identified different precipitators such as rainfall and seasonal variations in river levels as the main precipitators contributing to floods in the aforesaid three major urban sites. 

We utilized CatBoost and XG Boost initially for the predictive modelling due to their highest accuracy in predicting rainfall. But due to the long duration of the epoch of data collected from standardized sources such as NOAA and IMD, we had to use LSTM and RNN with gradient boosting for accurate predictions. 

We aim to publish this research first in an IEEE Conference, patent it and licence it to key stakeholders in the urban planning and disaster management domain.

The project intent included not only disaster mitigation and disaster preparedness but urban planning and town development utility as well